+++ src/main/java/org/mariadb/jdbc/internal/packet/send/SendHandshakeResponsePacket.java
@@ -49,11 +49,9 @@ OF SUCH DAMAGE.
 
 package org.mariadb.jdbc.internal.packet.send;
 
-import org.mariadb.jdbc.Driver;
 import org.mariadb.jdbc.MariaDbDatabaseMetaData;
 import org.mariadb.jdbc.internal.MariaDbServerCapabilities;
 import org.mariadb.jdbc.internal.protocol.authentication.DefaultAuthenticationProvider;
-import org.mariadb.jdbc.internal.util.Options;
 import org.mariadb.jdbc.internal.util.Utils;
 import org.mariadb.jdbc.internal.stream.PacketOutputStream;
 import org.mariadb.jdbc.internal.util.constant.Version;
@@ -62,7 +60,6 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.io.UnsupportedEncodingException;
 import java.lang.management.ManagementFactory;
-import java.nio.ByteBuffer;
 import java.security.NoSuchAlgorithmException;
 import java.util.Arrays;
 import java.util.StringTokenizer;
@@ -202,7 +199,11 @@ public class SendHandshakeResponsePacket implements InterfaceSendPacket {
         writeStringLength("_client_name", MariaDbDatabaseMetaData.DRIVER_NAME);
         writeStringLength("_client_version", Version.version);
         writeStringLength("_os", System.getProperty("os.name"));
-        writeStringLength("_pid", ManagementFactory.getRuntimeMXBean().getName());
+        try {
+            writeStringLength("_pid", ManagementFactory.getRuntimeMXBean().getName());
+        } catch (NoClassDefFoundError e) {
+            /* [CONJ-281] GAE support */
+        }
         writeStringLength("_thread", Long.toString(serverThreadId));
         writeStringLength("_java_vendor", System.getProperty("java.vendor"));
         writeStringLength("_java_version", System.getProperty("java.version"));