+++ documentation/Changelog.md
@@ -11,6 +11,7 @@ Not released
 * [CONJ-284] Cannot read autoincremented IDs bigger than Short.MAX_VALUE
 * [CONJ-283] Parsing correction on MariaDbClientPreparedStatement - syntax error on insert values
 * [CONJ-282] Handling YEARs with binary prepareStatement
+* [CONJ-281] Connector/J is incompatible with Google App Engine correction
 * [CONJ-278] Improve prepared statement on failover
 
 ## 1.4.2