+++ documentation/Changelog.md
@@ -7,7 +7,7 @@
 
 ---
 ## 1.4.3
-Not released 
+Released on 22 april 2016 
 * [CONJ-284] Cannot read autoincremented IDs bigger than Short.MAX_VALUE
 * [CONJ-283] Parsing correction on MariaDbClientPreparedStatement - syntax error on insert values
 * [CONJ-282] Handling YEARs with binary prepareStatement